Add test that triggers sanitizer error for the munmap_chunk error
This is a more complex traverse_postorder_with_stack operation that involves
traversing a bigger tree. This finally triggers the error sometimes seen in the
wild. It will require some debugging, but making the error reproducable was the
first big step.

The issue is in line 103 of op_manager.hpp.